+++ videoclub/src/main/java/formation/sopra/repository/FilmRepository.java
@@ -1,11 +1,11 @@
 package formation.sopra.repository;
 
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import formation.sopra.entity.Film;
+
 @Repository
-public interface FilmRepository extends JpaRepository<Film, Integer> {
+public interface FilmRepository extends JpaRepository<Film, Long> {
 
 }
